fix : profileType 폴더 변경
